Refactor model offset
